Made locking on Schema more fine-grained.

+++ src/main/kotlin/ch/unibas/dmi/dbis/cottontail/database/schema/Schema.kt
@@ -2,30 +2,30 @@ package ch.unibas.dmi.dbis.cottontail.database.schema
 
 import ch.unibas.dmi.dbis.cottontail.database.catalogue.Catalogue
 import ch.unibas.dmi.dbis.cottontail.database.column.mapdb.MapDBColumn
-import ch.unibas.dmi.dbis.cottontail.database.general.DBO
-import ch.unibas.dmi.dbis.cottontail.model.basics.ColumnDef
 import ch.unibas.dmi.dbis.cottontail.database.entity.Entity
 import ch.unibas.dmi.dbis.cottontail.database.entity.EntityHeader
 import ch.unibas.dmi.dbis.cottontail.database.entity.EntityHeaderSerializer
+import ch.unibas.dmi.dbis.cottontail.database.general.DBO
+import ch.unibas.dmi.dbis.cottontail.model.basics.ColumnDef
 import ch.unibas.dmi.dbis.cottontail.model.exceptions.DatabaseException
-import ch.unibas.dmi.dbis.cottontail.utilities.name.*
-
-import org.mapdb.*
-import org.mapdb.volume.MappedFileVol
+import ch.unibas.dmi.dbis.cottontail.utilities.extensions.read
+import ch.unibas.dmi.dbis.cottontail.utilities.extensions.write
+import ch.unibas.dmi.dbis.cottontail.utilities.name.Name
+import ch.unibas.dmi.dbis.cottontail.utilities.name.NameType
+import org.mapdb.DBException
+import org.mapdb.Serializer
+import org.mapdb.Store
+import org.mapdb.StoreWAL
 import java.io.IOException
-import java.lang.IllegalArgumentException
-
 import java.lang.ref.SoftReference
 import java.nio.file.Files
-
 import java.nio.file.Path
-
 import java.util.*
-import java.util.concurrent.locks.ReentrantReadWriteLock
+import java.util.concurrent.ConcurrentHashMap
+import java.util.concurrent.locks.ReentrantLock
+import java.util.concurrent.locks.StampedLock
 import java.util.stream.Collectors
-
-import kotlin.concurrent.read
-import kotlin.concurrent.write
+import kotlin.concurrent.withLock
 
 /**
  * Represents an schema in the Cottontail DB data model. A [Schema] is a collection of [Entity] objects that belong together
@@ -39,9 +39,21 @@ import kotlin.concurrent.write
  * @see MapDBColumn
  *
  * @author Ralph Gasser
- * @version 1.1
+ * @version 1.2
  */
 class Schema(override val name: Name, override val path: Path, override val parent: Catalogue) : DBO {
+
+    /**
+     * Companion object with different constants.
+     */
+    companion object {
+        /** ID of the schema header! */
+        const val HEADER_RECORD_ID: Long = 1L
+
+        /** Filename for the [Schema] catalogue.  */
+        const val FILE_CATALOGUE = "index.db"
+    }
+
     /** Constant FQN of the [Schema] object. */
     override val fqn: Name = this.parent.fqn.append(this.name)
 
@@ -54,23 +66,31 @@ class Schema(override val name: Name, override val path: Path, override val pare
 
     /** Reference to the [SchemaHeader] of the [Schema]. */
     private val header
-            get() = this.store.get(HEADER_RECORD_ID, SchemaHeaderSerializer) ?: throw DatabaseException.DataCorruptionException("Failed to open header of schema $fqn!")
+        get() = this.store.get(HEADER_RECORD_ID, SchemaHeaderSerializer)
+                ?: throw DatabaseException.DataCorruptionException("Failed to open header of schema $fqn!")
 
-    /** A lock used to mediate access to this [Schema]. */
-    private val lock = ReentrantReadWriteLock()
+    /** A lock used to mediate access the closed state of this [Schema]. */
+    private val closeLock = StampedLock()
+
+    /** A lock used to mediate access to changes to the entities contained in this [Schema]. */
+    private val entityLock = StampedLock()
+
+    /** A lock used to mediate access to [loaded] cache. */
+    private val cacheLock = ReentrantLock()
 
     /** A map of loaded [Entity] references. The [SoftReference] allow for coping with changing memory conditions. */
-    private val loaded = HashMap<Name, SoftReference<Entity>>()
+    private val loaded = ConcurrentHashMap<Name, SoftReference<Entity>>()
 
     /** Returns a list of [Entity] held by this [Schema]. */
     val entities: List<Name>
-        get() = this.header.entities.map { Name(this.store.get(it, Serializer.STRING) ?: throw DatabaseException.DataCorruptionException("Failed to read schema $fqn ($path): Could not find entity name of ID $it.")) }
-
-
+        get() = this.header.entities.map {
+            Name(this.store.get(it, Serializer.STRING)
+                    ?: throw DatabaseException.DataCorruptionException("Failed to read schema $fqn ($path): Could not find entity name of ID $it."))
+        }
 
     /** Size of the [Schema] in terms of [Entity] objects it contains. */
     val size
-        get() = this.lock.read { this.header.entities.size }
+        get() = this.entityLock.read { this.header.entities.size }
 
     /** Flag indicating whether or not this [Schema] has been closed. */
     @Volatile
@@ -82,58 +102,65 @@ class Schema(override val name: Name, override val path: Path, override val pare
      *
      * @param name The name of the [Entity] that should be created.
      */
-    fun createEntity(name: Name, vararg columns: ColumnDef<*>) = this.lock.write {
+    fun createEntity(name: Name, vararg columns: ColumnDef<*>) = this.closeLock.read {
+        /* Check closed status. */
+        if (this.closed) {
+            throw IllegalStateException("Schema ${this.name} has been closed and cannot be used anymore.")
+        }
+
         /* Check the type of name. */
         if (name.type != NameType.SIMPLE) {
             throw IllegalArgumentException("The provided name '$name' is of type '${name.type} and cannot be used to access an entity through a schema.")
         }
 
-        if (this.entities.contains(name)) throw DatabaseException.EntityAlreadyExistsException(this.fqn.append(name))
         if (columns.map { it.name }.distinct().size != columns.size) throw DatabaseException.DuplicateColumnException(this.fqn.append(name), columns.map { it.name })
+        if (this.entities.contains(name)) throw DatabaseException.EntityAlreadyExistsException(this.fqn.append(name))
 
-        /* Create empty folder for entity. */
-        val data = path.resolve("entity_$name")
+        this.entityLock.write {
+            /* Create empty folder for entity. */
+            val data = path.resolve("entity_$name")
 
-        try {
-            if (!Files.exists(data)) {
-                Files.createDirectories(data)
-            } else {
-                throw DatabaseException("Failed to create entity '${this.fqn.append(name)}'. Data directory '$data' seems to be occupied.")
+            try {
+                if (!Files.exists(data)) {
+                    Files.createDirectories(data)
+                } else {
+                    throw DatabaseException("Failed to create entity '${this.fqn.append(name)}'. Data directory '$data' seems to be occupied.")
+                }
+
+                /* Store entry for new entity. */
+                val recId = this.store.put(name.name, Serializer.STRING)
+
+                /* Generate the entity. */
+                val volumeFactory = this.parent.config.volumeFactory
+                val store = StoreWAL.make(file = data.resolve(Entity.FILE_CATALOGUE).toString(), volumeFactory = volumeFactory, fileLockWait = this.parent.config.lockTimeout)
+                store.preallocate() /* Pre-allocates the header. */
+
+                /* Initialize the entities header. */
+                val columnIds = columns.map {
+                    MapDBColumn.initialize(it, data, volumeFactory)
+                    store.put(it.name.name, Serializer.STRING)
+                }.toLongArray()
+                store.update(Entity.HEADER_RECORD_ID, EntityHeader(columns = columnIds), EntityHeaderSerializer)
+                store.commit()
+                store.close()
+
+                /* Update schema header. */
+                val header = this.header
+                header.modified = System.currentTimeMillis()
+                header.entities = header.entities.copyOf(header.entities.size + 1)
+                header.entities[header.entities.size - 1] = recId
+                this.store.update(HEADER_RECORD_ID, header, SchemaHeaderSerializer)
+
+                /* Commit changes to local schema. */
+                this.store.commit()
+            } catch (e: DBException) {
+                this.store.rollback()
+                val pathsToDelete = Files.walk(data).sorted(Comparator.reverseOrder()).collect(Collectors.toList())
+                pathsToDelete.forEach { Files.delete(it) }
+                throw DatabaseException("Failed to create entity '${this.fqn.append(name)}' due to error in the underlying data store: {${e.message}")
+            } catch (e: IOException) {
+                throw DatabaseException("Failed to create entity '${this.fqn.append(name)}' due to an IO exception: {${e.message}")
             }
-
-            /* Store entry for new entity. */
-            val recId = this.store.put(name.name, Serializer.STRING)
-
-            /* Generate the entity. */
-            val volumeFactory = this.parent.config.volumeFactory;
-            val store = StoreWAL.make(file = data.resolve(Entity.FILE_CATALOGUE).toString(), volumeFactory = volumeFactory, fileLockWait = this.parent.config.lockTimeout)
-            store.preallocate() /* Pre-allocates the header. */
-
-            /* Initialize the entities header. */
-            val columnIds = columns.map {
-                MapDBColumn.initialize(it, data, volumeFactory)
-                store.put(it.name.name, Serializer.STRING)
-            }.toLongArray()
-            store.update(Entity.HEADER_RECORD_ID, EntityHeader(columns = columnIds), EntityHeaderSerializer)
-            store.commit()
-            store.close()
-
-            /* Update schema header. */
-            val header = this.header
-            header.modified = System.currentTimeMillis()
-            header.entities = header.entities.copyOf(header.entities.size + 1)
-            header.entities[header.entities.size-1] = recId
-            this.store.update(HEADER_RECORD_ID, header, SchemaHeaderSerializer)
-
-            /* Commit changes to local schema. */
-            this.store.commit()
-        } catch (e: DBException) {
-            this.store.rollback()
-            val pathsToDelete = Files.walk(data).sorted(Comparator.reverseOrder()).collect(Collectors.toList())
-            pathsToDelete.forEach { Files.delete(it) }
-            throw DatabaseException("Failed to create entity '${this.fqn.append(name)}' due to error in the underlying data store: {${e.message}")
-        } catch (e: IOException) {
-            throw DatabaseException("Failed to create entity '${this.fqn.append(name)}' due to an IO exception: {${e.message}")
         }
     }
 
@@ -142,41 +169,49 @@ class Schema(override val name: Name, override val path: Path, override val pare
      *
      * @param name The name of the [Entity] that should be dropped.
      */
-    fun dropEntity(name: Name) = this.lock.write {
+    fun dropEntity(name: Name) = this.closeLock.read {
+        /* Check closed status. */
+        if (this.closed) {
+            throw IllegalStateException("Schema ${this.name} has been closed and cannot be used anymore.")
+        }
+
         /* Check the type of name. */
         if (name.type != NameType.SIMPLE) {
             throw IllegalArgumentException("The provided name '$name' is of type '${name.type} and cannot be used to access an entity through a schema.")
         }
 
-        val entityRecId = this.header.entities.find { this.store.get(it, Serializer.STRING) == name.name } ?: throw DatabaseException.EntityDoesNotExistException(fqn.append(name))
+        this.entityLock.write {
+            val entityRecId = this.header.entities.find { this.store.get(it, Serializer.STRING) == name.name }
+                    ?: throw DatabaseException.EntityDoesNotExistException(fqn.append(name))
 
-        val entity = this.entityForName(name)
-        entity.allIndexes().map{ it.name }.forEach { entity.dropIndex(it) }
+            val entity = this.entityForName(name)
+            entity.allIndexes().map { it.name }.forEach { entity.dropIndex(it) }
 
-        /* Unload the entity and remove it. */
-        this.unload(name)
+            /* Unload the entity and remove it. */
+            this.loaded.remove(name)?.get()?.close()
 
-        /* Remove entity. */
-        try {
-            /* Delete entity name from list. */
-            this.store.delete(entityRecId, Serializer.STRING)
+            /* Remove entity. */
+            try {
+                /* Delete entity name from list. */
+                this.store.delete(entityRecId, Serializer.STRING)
 
-            /* Update header. */
-            val header = this.header
-            header.modified = System.currentTimeMillis()
-            header.entities = header.entities.filter { it != entityRecId }.toLongArray()
-            this.store.update(HEADER_RECORD_ID, header, SchemaHeaderSerializer)
+                /* Update header. */
+                val header = this.header
+                header.modified = System.currentTimeMillis()
+                header.entities = header.entities.filter { it != entityRecId }.toLongArray()
+                this.store.update(HEADER_RECORD_ID, header, SchemaHeaderSerializer)
 
-            /* Commit. */
-            this.store.commit()
-        } catch (e: DBException) {
-            this.store.rollback()
-            throw DatabaseException("Entity '${this.fqn.append(name)}' could not be dropped, because of an error in the underlying data store: ${e.message}!")
+                /* Commit. */
+                this.store.commit()
+            } catch (e: DBException) {
+                this.store.rollback()
+                throw DatabaseException("Entity '${this.fqn.append(name)}' could not be dropped, because of an error in the underlying data store: ${e.message}!")
+            }
+
+            /* Delete all files associated with the entity. */
+            val pathsToDelete = Files.walk(this.path.resolve("entity_$name")).sorted(Comparator.reverseOrder()).collect(Collectors.toList())
+            pathsToDelete.forEach { Files.deleteIfExists(it) }
         }
-
-        /* Delete all files associated with the entity. */
-        val pathsToDelete = Files.walk(this.path.resolve("entity_$name")).sorted(Comparator.reverseOrder()).collect(Collectors.toList())
-        pathsToDelete.forEach { Files.deleteIfExists(it) }
     }
 
 
@@ -186,15 +221,29 @@ class Schema(override val name: Name, override val path: Path, override val pare
      *
      * @param name Name of the [Entity] to access.
      */
-    fun entityForName(name: Name): Entity = this.lock.read {
+    fun entityForName(name: Name): Entity = this.closeLock.read {
+        /* Check closed status. */
+        if (this.closed) {
+            throw IllegalStateException("Schema ${this.name} has been closed and cannot be used anymore.")
+        }
+
         /* Check the type of name. */
         if (name.type != NameType.SIMPLE) {
             throw IllegalArgumentException("The provided name '$name' is of type '${name.type}  and cannot be used to access an entity through a schema.")
         }
 
-        if (!this.entities.contains(name)) throw DatabaseException.EntityDoesNotExistException(this.fqn.append(name))
-        return this.loaded[name]?.get() ?: Entity(name, this).also {
-            this.loaded[name] = SoftReference(it)
+        this.entityLock.read {
+            if (!this.entities.contains(name)) throw DatabaseException.EntityDoesNotExistException(this.fqn.append(name))
+            this.cacheLock.withLock {
+                var ret = this.loaded[name]?.get()
+                return if (ret != null) {
+                    ret
+                } else {
+                    ret = Entity(name, this)
+                    this.loaded[name] = SoftReference(ret)
+                    ret
+                }
+            }
         }
     }
 
@@ -205,33 +254,17 @@ class Schema(override val name: Name, override val path: Path, override val pare
      * objects may be held by other threads, it can take a
      * while for this method to complete.
      */
-    override fun close() = this.lock.write {
+    override fun close() = this.closeLock.read {
         if (!this.closed) {
-            this.loaded.entries.removeIf {
-                it.value.get()?.close()
-                true
+            this.cacheLock.withLock {
+                this.loaded.entries.removeIf {
+                    it.value.get()?.close()
+                    true
+                }
             }
             this.store.close()
+            this.closed = true
         }
-    }
-
-    /**
-     * Explicitly unloads (closes) an instance of [Entity] if such an instance exists. Otherwise, this function has now effect.
-     */
-    private fun unload(name: Name) = this.lock.write {
-        this.loaded[name]?.get()?.close() /* BLOCK! */
-        this.loaded.remove(name)
-    }
-
-    /**
-     * Companion object with different constants.
-     */
-    companion object {
-        /** ID of the schema header! */
-        const val HEADER_RECORD_ID: Long = 1L
-
-        /** Filename for the [Schema] catalogue.  */
-        const val FILE_CATALOGUE = "index.db"
     }
 }
